In-cluster eventing flow without check of Pod logs
