+++ tests/fixtures/in-cluster-eventing-receiver-dependencies.json
@@ -1,0 +1,5 @@
+{
+  "name": "in-cluster-eventing-receiver",
+  "version": "1.0.0",
+  "dependencies": {}
+}
