+++ tests/support/commands.js
@@ -67,3 +67,7 @@ function paste(subject, { pastePayload }) {
 
   return subject;
 }
+
+Cypress.Commands.add('getLeftNav', () => {
+  cy.get('nav[data-testid=semiCollapsibleLeftNav]');
+});
