+++ tests/support/index.js
@@ -1,3 +1,5 @@
 import './commands';
 import './iframe-commands';
 import './login';
+import './views/api-rules-view';
+import './views/functions-view';
